# docs/biz-reporting-openapi-initial.yaml
openapi: 3.0.3
info:
  title: Business Unit Reporting API
  version: 0.1.0
  description: Initial OpenAPI draft for Workbuddy agent implementation.
servers:
  - url: https://api.example.com
...
paths:
...
  /api/city/packages/current:
    get:
      tags: [Packages]
      summary: Get or initialize current annual package
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: year
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
  /api/city/packages/{packageId}/months/{monthNo}:
    get:
      tags: [Packages]
      summary: Get editable month data
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: packageId
          required: true
          schema:
            type: integer
        - in: path
          name: monthNo
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
  /api/city/packages/{packageId}/draft-save:
    post:
      tags: [Packages]
      summary: Save month draft
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: packageId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftSaveRequest'
      responses:
        '200':
          description: Saved
  /api/city/packages/{packageId}/submit-preview:
    post:
      tags: [Packages]
      summary: Submit preview
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: packageId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftSaveRequest'
      responses:
        '200':
          description: Preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitPreviewResponse'
  /api/city/packages/{packageId}/submit:
    post:
      tags: [Packages]
      summary: Submit month
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: packageId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftSaveRequest'
      responses:
        '200':
          description: Submitted
  /api/city/packages/{packageId}/read-only-months/{monthNo}:
    get:
      tags: [Snapshots]
      summary: Read-only month detail
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: packageId
          required: true
          schema:
            type: integer
        - in: path
          name: monthNo
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
  /api/admin/packages/{packageId}/return-to-draft:
    post:
      tags: [AdminPackages]
      summary: Return submitted month to draft
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: packageId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnToDraftRequest'
      responses:
        '200':
          description: Updated
  /api/admin/packages/{packageId}/unlock-months:
    post:
      tags: [AdminPackages]
      summary: Unlock historical months
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: packageId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnlockMonthsRequest'
      responses:
        '200':
          description: Unlocked
  /api/admin/packages/{packageId}/open-current-month-contract:
    post:
      tags: [AdminPackages]
      summary: Allow current-month new contract fill
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: packageId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenCurrentMonthContractRequest'
      responses:
        '200':
          description: Opened
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ContractMonthInput:
      type: object
      required: [contractId, completionAmount, acceptanceAmount]
      properties:
        contractId:
          type: integer
        completionAmount:
          type: number
        acceptanceAmount:
          type: number
    CostMonthInput:
      type: object
      required: [costCategoryCode, amount]
      properties:
        costCategoryCode:
          type: string
        amount:
          type: number
    MaintenanceMonthInput:
      type: object
      properties:
        invoiceTotalPrevYear:
          type: number
        invoiceMonthCountPrevYear:
          type: integer
        invoiceTotalCurrentYear:
          type: number
    DraftSaveRequest:
      type: object
      required: [monthNo, contractRows, costRows]
      properties:
        monthNo:
          type: integer
        contractRows:
          type: array
          items:
            $ref: '#/components/schemas/ContractMonthInput'
        costRows:
          type: array
          items:
            $ref: '#/components/schemas/CostMonthInput'
        maintenanceRow:
          $ref: '#/components/schemas/MaintenanceMonthInput'
...
    SubmitPreviewResponse:
      type: object
      properties:
        belongMonth:
          type: integer
        completionTotal:
          type: number
        acceptanceTotal:
          type: number
        costTotal:
          type: number
        grossProfit:
          type: number
        isOverdue:
          type: boolean
...
    ReturnToDraftRequest:
      type: object
      required: [monthNo, reason]
      properties:
        monthNo:
          type: integer
        reason:
          type: string
    UnlockMonthsRequest:
      type: object
      required: [months, expiresAt]
      properties:
        months:
          type: array
          items:
            type: integer
        expiresAt:
          type: string
          format: date-time
        reason:
          type: string
          nullable: true
    OpenCurrentMonthContractRequest:
      type: object
      required: [monthNo, contractId]
      properties:
        monthNo:
          type: integer
        contractId:
          type: integer
...